Add Osuuspankki-elasku receipt scripts
oth scripts are needed. saveoplasku.sh is a script which calls saveOPlasku.rec.
Usage: saveoplasku <CompanyBillIsFrom> <DatePaymentWasMade>  (Firefox as browser, open to OP pages and having opened an elasku that we want a receipt from, open a terminal window and run the script.)
Example: saveoplasku Securitas 2017_05_03 
Format of date is not significant, it is just added as a string to the end of the bill name to help distinguish between bills paid to the same company.
saveOPlasku.rec is a keyboard macro which just spits out certain key strokes. These are *timed*, but due to internet conection or computer lag time may not work correctly 100% of the time. If that happens, Ctrl-C can interrupt the macro, but needs to be done from the terminal the script was started from. (Alt-TAB ONCE to get back to the previous window, then Ctrl-C)