Log every 10 seconds (not 1), add better error msg
When logging every second, there was a pretty high probability that shutdown would occur during writing to the log, resulting in garbage on the last line of the log, which then renders it somewhat "unreadable" by this script later. Logging every 10 seconds lowers that probability significantly AND we still get logs which are sufficient.

Also added a slightly better error message in the event that there is garbage at the end of the log file and it needs to be removed.